vendor/mpir_pmi: Add MPIR_pmi_barrier_group to sync across a group of procs.
Cherry-pick partly upstream commit 8f6801f. The addition of the interface
MPIR_pmi_allgather_group is omitted because it is not required for now.

For simplicity, the upstream commits adding group barrier support to PMI-1
are not added here. This can be done later or with vendor merge of MPICH 5.0.