vendor/comm: Enable release of uncommitted communicator.
This commit enables to report errors that happen before or during
communicator commit to the upper layers by enabling the release of
a partially initialized/ committed communicator. Such cases can
happen, for example, when there are errors in the pre commit hook
of the device.

The following changes are applied:
- Protect coll cleanup against uninitialized collective settings
- No message check in internal delete for uncommitted comm
- Do not free temporary session context id on comm delete