mpix/test: Add test for MPIX_Info_split_into_array().
This test splits an initially merged info object back into an array
of info objects by using `MPIX_Info_split_into_array()` and checks
for the correct behavior if different sizes for this array are used.

In addition, the test also checks if cases where more info objects are
given in the array than actually needed are still handled correctly.
Finally, it also checks what happens if no objects are given in the
array.