init/comm: Move connection init to comm pre-commit hook.
Postpone the connection init to the communicator creation by moving it
from MPID_Init to MPIDI_PSP_Comm_commit_pre_hook. This way, MPI_Session_init
is a local procedure that does not have any collective operation involved
such as a barrier over the KVS.

As a result, partial connectivity is possible if the first created comm is
not the world communicator. Hence, some changes were required to deal with
the fact that an MPI application may only reach partial connectivity over
its lifetime and some connections may not exist at all at the end, i.e.,
must not be disconnected.