vendor/test: Add test session/session_4x4.
Backport upstream commit 11895ac. This commit will be included
as of MPICH 5.0.0 and is part of MPICH's true MPI Session
implementation. In addition, upstream commit d47dc02 including
clean ups is backported.

Add a test that calls MPI_Comm_create_from_group from a group
that is not an original pset.

Differences to upstream: The new test runs only if PMIx is used.