psp/compr: Add MPIX compressor support to PSP device.
The implementation of the `MPIX_Register_compressor()` functionality
is envisaged on several levels: While the previous commit implements
the pure registration of the callback functions, the actual invocation
of the compression and decompression functions is foreseen at the ADI3
level.

This commit implements this for the PSP device and its realization of
partitioned communication through checking the corresponding info object
for the key "compressor". If a matching compressor is found, the PSP
device transparently invokes the registered callback functions:
- `req_init_fn()` during `MPI_Psend_init()`
- `deflate_fn()` during `MPI_Pready()`
- `inflate_fn()` when message arrived (pscom `io_done` callback)
- `req_free_fn()` during `MPI_Request_free()`

In addition, this commit also implements the transparent use of compressor
plugins (as envisaged in the proposal) within the PSP device by searching
for a suitable plugin if no matching compressor was found in the list of
the already registered ones -- provided that the name of such a plugin
has been specified via the info key "compressor_plugin" and that this
mechanism has been enabled explicitly by setting the environment variable
`PSP_COMPRESSOR_PLUGINS=1`.