vendor/compr: Add MPIX_Register_compressor().
This commit is the result of the efforts of the DaREXA-F project to
create an MPI extension that enables flexible and transparent data
compression of MPI payloads, while allowing also for a conscious
control of the respective parameters, for example, in case of lossy
compression. At the end, these efforts boiled down to adding a single
new MPI function that enables compression algorithms to be registered
in the form of callback functions to be invoked transparently:

`MPIX_Register_compressor(name, req_init_fn, req_free_fn, deflate_fn,
                          inflate_fn, deregister_fn, info, extra_state)`

In addition to the compression/decompression functions, callbacks for
buffer handling can also be registered, and corresponding wildcards
are provided if some functions are not to be implemented individually.

All further control is then handled via info objects and corresponding
key/value pairs, while particular support is intended for partitioned
communication, whose init functions for sending/receiving also foresee
the passing of an info object.