vendor/compr: Return compressor list via MPI_Info_create_env().
Since compressors do not have a handle, there are also no compressor-
related counterparts to the common info set/get functions for MPI
objects. However, an application should be able to query at least
general compressor-related information as, for instance, the list of
names of all registered compressors -- especially since these may also
be pre-registered.

For this purpose, this commit attaches this information to the info
objects returned by `MPI_Info_create_env()` in two ways:
(1) as an array-type info entry (with the key "compressor"), and
(2) as a comma-separated list (with the key "compressor_list").