test/compr: Add test for compressor info via MPI_Info_create_env().
This adds a new test program that checks the availability of compressor
related keys in the info object returned by `MPI_Info_create_env()`.

The test registers two compressors via `MPIX_Register_compressor()`
first and then calls `MPI_Info_create_env()`.
According to the proposed behavior, it is then checked (1) whether
there is an array-type entry with two elements for the key "compressor"
(each of which corresponding to one of the registered compressor names)
and (2) if there is also an entry for the key "compressor_list" that
features a comma-separated string representing a list of the compressor
names, with an order that matches the one of the array-type entry.