vendor/ucc: Add mechanics for data type packing.
In the event that a sender uses basic data types in a collective
operation, but a receiver uses derived data types (or vice versa),
a fallback of the UCC wrapper to MPICH's collectives cannot be done
consistently -- resulting in deadlocks. (See, for example, `gather2`
and `scatterv` in `test/mpi/coll`.)

This commit attempts to fix this by trying to break down derived data
types to their basic data types, which in the end must match the type
map on the other side, since otherwise the application would be faulty.

However, this approach can have the disadvantage of causing additional
copy operations. The advantage, on the other hand, is that it can be
assumed that both sides are acting consistently and that no deadlock
will occur in such mixed cases.

(Please note that the case described above cannot occur for reduction
operations, since neither derived datatypes nor specifying different
datatypes for the senders and the receiver are here permitted.)